requirement.txt
numpy
opencv-python
torch
tensorflow
pillow
torchvision
collection